move popover types
