+++ types/configs/index.d.ts
@@ -1,3 +1,5 @@
+import { fromCallback } from 'cypress/types/bluebird';
+
 export * from './editor-config';
 export * from './sanitizer-config';
 export * from './paste-config';
@@ -5,3 +7,4 @@ export * from './conversion-config';
 export * from './log-levels';
 export * from './i18n-config';
 export * from './i18n-dictionary';
+export * from './popover'
